refactor(persistence): move large inline test modules into module-local files (TD-026)
Relocate the big #[cfg(test)] blocks out of the production persistence
source files into module-local test files so the production source is
easier to scan. Test-relocation only — no assertions changed.

- database/mod.rs (853-line inline tests) → database/tests.rs
- lock.rs converted to lock/ directory module; tests (307 lines) → lock/tests.rs

cargo test -p parish-persistence: 127 passed (same tests, new location).
architecture_fitness: 5 passed (no orphaned files).
Closes #1202 (TD-026).
